Refactored: plg_privacy_kunena
