+++ src/plugins/plg_kunena_uddeim/src/Extension/Uddeim.php
@@ -69,7 +69,7 @@ class Easyblog extends CMSPlugin implements SubscriberInterface, DatabaseAwareIn
     }
 
     /**
-     * plgKunenaEasyblog constructor.
+     * plgKunenaUddeim constructor.
      *
      * @param  array $config
      */
