+++ src/plugins/plg_privacy_kunena/services/provider.php
@@ -35,15 +35,15 @@ return new class implements ServiceProviderInterface
     public function register(Container $container)
     {
         $container->set(
-        	PluginInterface::class,
-        	function (Container $container) {
+            PluginInterface::class,
+            function (Container $container) {
                 $subject = $container->get(DispatcherInterface::class);
                 $config  = (array) PluginHelper::getPlugin('privacy', 'kunena');
-                $plugin = new Kunena(
-                	$subject,
-                	$config
-                );
+                $plugin  = new Kunena($config);
+
+                $plugin->traitSetDispatcher($subject);
                 $plugin->setApplication(Factory::getApplication());
+                $plugin->setDatabase($container->get('DatabaseDriver'));
 
                 return $plugin;
             }
